Mod geometries (#14)
* work on deltatime

* added include for sorting

* fixed compile issues

* begin impliment AnaSummary

* allow multiple histos of the same run number to be compared

* updated CalibSummary and labels

* debug

* added module numbers

* added input for module positions

* add modpos to output map

* impliment module positions

* updated module position

* minor change

* updated for new CreateMapping macro

* added module dependence

* updated with new module positions

* updated HGCROC header

* New mapping file CAEN

* updated mapping files

* Update convertData_2024.sh

---------

+++ NewStructure/Event.h
@@ -28,7 +28,7 @@ class Event{
   int     GetRunNumber (void) const;
   Tile*   GetTile      (int);
   Tile*   GetTileFromID(int);
-  int     GetTimeStamp (void) const;
+  double     GetTimeStamp (void) const;
   double  GetVov       (void) const;
   double  GetVop       (void) const;
   double    GetBeamPosX(void);
@@ -47,7 +47,7 @@ class Event{
   void    SetVop(double);
   void    SetBeamPosX(double);
   void    SetBeamPosY(double);
-  void    SetTimeStamp (int);
+  void    SetTimeStamp (double);
 
   bool    InspectIfLocalMuonTrigg(int, double, double, double);
   bool    InspectIfNoiseTrigg(int, double, double);
@@ -68,7 +68,7 @@ class Event{
   double                BeamPosY;
   std::map<int, Tile* > Tiles;
   std::vector<int>      TileIDs;
-  int                   TimeStamp;
+  double                   TimeStamp;
   
  protected:
